requirements.txt
openai
python-dotenv
langchain-openai
langchain-ollama
google
gradio
langchain_community
faiss-cpu
anthropic
langchain_anthropic
langchain_huggingface
transformers
boto3
streamlit
langchain_aws
pypdf
chardet
fastapi
uvicorn
#GmailLoader
#unstructured[all-docs]
networkx
msoffcrypto-tool
openpyxl
tools